Little changes

+++ Bachelor_app/Configuration.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using Bachelor_app.Helper;
+using System;
+using System.IO;
 
 namespace Bachelor_app
 {
@@ -12,10 +14,12 @@ namespace Bachelor_app
         public static string TempDrawKeypoint { get; private set; } = Path.Combine(Configuration.TempDirectoryPath, "DrawKeypoint");
         public static string TempDrawMatches { get; private set; } = Path.Combine(Configuration.TempDirectoryPath, "DrawMatch");
         #endregion
+
         #region Match file
         public static string MatchFileName { get; private set; } = "AllFoundedMatches.txt";
         public static string MatchFilePath { get; private set; } = Path.Combine(TempDirectoryPath, MatchFileName);
         #endregion
+
         #region Tools
         public static string ToolsPath { get; private set; } = Path.GetFullPath(".\\Tools");
         public static string VisualSFMTool { get; private set; } = "VisualSFM.exe";
@@ -24,6 +28,7 @@ namespace Bachelor_app
         public static string VisualSFMResult { get; private set; } = "Result.nvm";
         public static string VisualSFMResultPath { get; private set; } = Path.Combine(Configuration.TempDirectoryPath, VisualSFMResult);
         #endregion
+
         #region Calibration
         public static string CalibrationName { get; private set; } = "Calibration.json";
         public static string CalibrationPath { get; private set; } = Path.Combine(TempDirectoryPath, CalibrationName);
@@ -39,32 +44,49 @@ namespace Bachelor_app
         //public static ListViewModel ListViewModel { get; private set; } = new ListViewModel();
         #endregion
 
-        public static void GenerateFolders()
+        public static void GenerateFolders(bool TryingAgain = false)
         {
-            //if (Directory.Exists(TempDirectoryPath))
-            //    DeleteTempFolder();
-
-            Directory.CreateDirectory(TempDirectoryPath);
-            Directory.CreateDirectory(TempDepthMapDirectoryPath);
-            Directory.CreateDirectory(TempLeftStackDirectoryPath);
-            Directory.CreateDirectory(TempRightStackDirectoryPath);
-            Directory.CreateDirectory(TempDrawKeypoint);
-            Directory.CreateDirectory(TempDrawMatches);
+            try
+            {
+                Directory.CreateDirectory(TempDirectoryPath);
+                Directory.CreateDirectory(TempDepthMapDirectoryPath);
+                Directory.CreateDirectory(TempLeftStackDirectoryPath);
+                Directory.CreateDirectory(TempRightStackDirectoryPath);
+                Directory.CreateDirectory(TempDrawKeypoint);
+                Directory.CreateDirectory(TempDrawMatches);
+            }
+            catch (Exception e)
+            {
+                if (!TryingAgain)
+                    GenerateFolders(true);
+                else
+                    throw e;
+            }
         }
 
         public static void DeleteTempFolder()
         {
-            DirectoryInfo di = new DirectoryInfo(TempDirectoryPath);
+            try
+            {
+                if (Directory.Exists(TempDirectoryPath))
+                {
+                    DirectoryInfo di = new DirectoryInfo(TempDirectoryPath);
 
-            foreach (FileInfo file in di.GetFiles())
-            {
-                file.Delete();
+                    foreach (FileInfo file in di.GetFiles())
+                    {
+                        file.Delete();
+                    }
+                    foreach (DirectoryInfo dir in di.GetDirectories())
+                    {
+                        dir.Delete(true);
+                    }
+                    di.Delete();
+                }
             }
-            foreach (DirectoryInfo dir in di.GetDirectories())
+            catch (Exception)
             {
-                dir.Delete(true);
+                WindowsFormHelper.AddLogToConsole($"Can't clear temp folder: {TempDirectoryPath}\n");
             }
-            di.Delete();
 
         }
     }
